\name{MPDiR-package}
\alias{MPDiR-package}
\alias{MPDiR}
\docType{package}
\title{
Modeling Psychophysical Data in R
}
\description{
The package contains data sets, functions and scripts for the book Modeling Psychophysical Data in R
}
\details{
\tabular{ll}{
Package: \tab MPDiR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-08-31\cr
License: \tab What license is it under?\cr
}

}
\author{
Kenneth Knoblauch and Laurence T. Maloney

Maintainer: Ken Knoblauch <ken.knoblauch@inserm.fr>
}
\references{
K. Knoblauch and L. T. Maloney, Modeling Psychophysical Data in R, Springer, (2012)
ISBN 978-1-4614-4475-6.
}

\keyword{ package }

